requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==4.2.20
django-filter==25.1
django-import-export==4.3.7
django-simple-history==3.8.0
djangorestframework==3.16.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.1
